Fix example of non-similar CSP sources with different ports
According to CSP match-ports algorithm
https://w3c.github.io/webappsec-csp/#match-ports, if a source
expression explicitly specify a port than only sources with the same
exact port will match. So the effective intersection of sources
matching both http://example.com:80/page1/html and
https://example.com:443/ is empty.

Consistently with that, the intersection of those two sources should
not be similar to each other and their intersection should be empty.